+++ activerecord/test/cases/tasks/mysql_rake_test.rb
@@ -219,44 +219,31 @@ module ActiveRecord
 
   class MySQLStructureDumpTest < ActiveRecord::TestCase
     def setup
-      @connection    = stub(:structure_dump => true)
       @configuration = {
         'adapter'  => 'mysql',
         'database' => 'test-db'
       }
-
-      ActiveRecord::Base.stubs(:connection).returns(@connection)
-      ActiveRecord::Base.stubs(:establish_connection).returns(true)
     end
 
     def test_structure_dump
       filename = "awesome-file.sql"
-      ActiveRecord::Base.expects(:establish_connection).with(@configuration)
-      @connection.expects(:structure_dump)
+      Kernel.expects(:system).with("mysqldump", "--result-file", filename, "--no-data", "test-db")
 
       ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, filename)
-      assert File.exists?(filename)
-    ensure
-      FileUtils.rm(filename)
     end
   end
 
   class MySQLStructureLoadTest < ActiveRecord::TestCase
     def setup
-      @connection    = stub
       @configuration = {
         'adapter'  => 'mysql',
         'database' => 'test-db'
       }
-
-      ActiveRecord::Base.stubs(:connection).returns(@connection)
-      ActiveRecord::Base.stubs(:establish_connection).returns(true)
-      Kernel.stubs(:system)
     end
 
     def test_structure_load
       filename = "awesome-file.sql"
-      Kernel.expects(:system).with('mysql', '--database', 'test-db', '--execute', %{SET FOREIGN_KEY_CHECKS = 0; SOURCE #{filename}; SET FOREIGN_KEY_CHECKS = 1})
+      Kernel.expects(:system).with('mysql', '--execute', %{SET FOREIGN_KEY_CHECKS = 0; SOURCE #{filename}; SET FOREIGN_KEY_CHECKS = 1}, "--database", "test-db")
 
       ActiveRecord::Tasks::DatabaseTasks.structure_load(@configuration, filename)
     end